pyproject.toml
[project]
name = "air-guardian"
version = "0.1.0"
description = "空气质量预测与智能分析系统"
authors = [
    {name = "AirGuardian Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "scikit-learn>=1.3.0",
    "requests>=2.31.0",
    "reportlab>=4.0.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
]

# 可选的加速支持
acceleration = [
    "GPUtil>=1.4.0",
    "nvidia-ml-py3>=7.352.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

